Implement cache invalidation

+++ lib/cucumber/website/git_hub/file_system_cache.rb
@@ -7,6 +7,7 @@ module Cucumber
       class FileSystemCache
         def initialize(api, config)
           @api = api
+          @cache_refresh_interval = config['git_hub']['cache_refresh_interval']
           @cache_path = Pathname.new("cache/git_hub/#{config['env']}.yaml")
         end
 
@@ -19,18 +20,22 @@ module Cucumber
           YAML::load(File.read(cache_path))
         end
 
-        attr_reader :api, :cache_path
-        private :api, :cache_path
+        attr_reader :api, :cache_path, :cache_refresh_interval
+        private :api, :cache_path, :cache_refresh_interval
 
         private
 
         def update_cache_if_needed
-          unless cache_exist?
+          unless cache_exist? && cache_valid?
             cache_path.dirname.mkpath
             File.open(cache_path, 'w+') do |file|
               file.write(YAML::dump(api.events))
             end
           end
+        end
+
+        def cache_valid?
+          File.mtime(cache_path) > Time.now - cache_refresh_interval
         end
 
         def cache_exist?
